+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # CHANGELOG
 
 ## 1.0 (unreleased)
-- CHANGED `--end` clause renamed from `--finally`
+- CHANGED:
+    - `--end` clause renamed from `--finally`
+    - `--text` renamed from `--whole`
 - auto-importing works in the `--setup` clause
 - will not internally change commands starting with a reserved keyword (ex: `if s ==` will not be changed to `s = if s ===`)
 - generator